% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-Checks.R
\name{CheckDBExists}
\alias{CheckDBExists}
\title{Checks if database exists.}
\usage{
CheckDBExists(shinyDir)
}
\arguments{
\item{shinyDir}{The path of the shiny App directory.}
}
\value{
TRUE if database exists
}
\description{
The demo shiny app relies on a SQLite database. This function checks whether this database exists or not
}
\details{
These are further details.
}
\section{A Custom Section}{


Text accompanying the custom section.
}

