% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-Checks.R
\name{CheckTblLLTunaBillfish}
\alias{CheckTblLLTunaBillfish}
\title{Checks the existence of Longline Data Table.}
\usage{
CheckTblLLTunaBillfish(shinyDir)
}
\arguments{
\item{shinyDir}{The path of the shiny App directory where the database is to be created.}
}
\value{
Creates a data base in the \code{shinyDir} directory. Throws an error if it fails.
}
\description{
Checks if the data containing longline data for tuna and billfish exists. This table contains data on longline fishing, with a grid of 5 degree spatial resolution, on a monthly base splitted by country.
}
