% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-Checks.R
\name{CreateDB}
\alias{CreateDB}
\title{Creates a database for the shiny app.}
\usage{
CreateDB(shinyDir)
}
\arguments{
\item{shinyDir}{The path of the shiny App directory where the database is to be created.}
}
\value{
Creates a data base in the \code{shinyDir} directory. Throws an error if it fails.
}
\description{
The demo shiny app relies on a SQLite database. This function creates an empty database.
}
