% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-Data.R
\name{GetFlagNames}
\alias{GetFlagNames}
\title{Get Flag names}
\usage{
GetFlagNames(shinyDir = ".")
}
\arguments{
\item{shinyDir}{The directory of the shiny app.}
}
\description{
Get a character vector with the abbreviations for the flags of the operating vessels.
}
\details{
Use this function to generate a list of valid choices of flags that are contained in the database.
}
