% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-Data.R
\name{GetSppNames}
\alias{GetSppNames}
\title{Get spp names}
\usage{
GetSppNames(shinyDir)
}
\arguments{
\item{shinyDir}{The directory of the shiny app.}
}
\description{
Get a character vector with the abbreviations for the species, plus a class 'ALL'.
}
\details{
Use this function to generate a list of valid choices of species that are contained in the database.
}
