% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-Data.R
\name{GetTable}
\alias{GetTable}
\title{Get a table from the database}
\usage{
GetTable(shinyDir = ".", table_name)
}
\arguments{
\item{shinyDir}{The directory of the shiny app.}

\item{table_name}{the \emph{descriptive name} of the table. See Details.}
}
\description{
Name the table you want, and get it.
}
\details{
\code{table_name} should have the following character values; otherwise returns an empty data frame:

\describe{
\item{General Spatial Data:}{Retrieves a table of catches by lon, lat and species. The database name for this table is \code{aux_summ_time_flag} (because data is summarized over time and flag.)}
\item{Catch Time Series:}{Retrieves a table with year, species and total weight, to plot a time series of catches.}
\item{Proportional Catches by Period-Flag:}{Contains catches scaled by the maximum catch within a time period. Also, contains the min and max years that composed a period for each flag.}
}
}
