% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-Data.R
\name{GetTimePeriod}
\alias{GetTimePeriod}
\title{Get a valid time period}
\usage{
GetTimePeriod(shinyDir = ".")
}
\arguments{
\item{shinyDir}{The directory of the shiny app.}
}
\description{
Get a integer vector of time periods that are contained in the database.
}
