% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-Map-Spp.R
\name{MapCatchesBySpp}
\alias{MapCatchesBySpp}
\title{Maps all catches by species}
\usage{
MapCatchesBySpp(catchData)
}
\arguments{
\item{catchData}{the reactive dataset filtered by the user.}
}
\description{
Gives a general spatial description of long-line fishing.
}
