% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotByFlag.R
\name{PlotByFlag}
\alias{PlotByFlag}
\title{Plot relative catches by flag and time period.}
\usage{
PlotByFlag(plotData)
}
\arguments{
\item{plotData}{A filtered table containing data from only one time period and one flag.
Data must come from \link{GetTable} using the value \emph{Proportional Catches by Period-Flag} for the argument \code{table_name}.}
}
\description{
Creates a barplot of catches scaled by maximum value, within a time period and by country flag..
}
