% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-Aux-Tables.R
\name{UpdateAuxiliaryTables}
\alias{UpdateAuxiliaryTables}
\title{Creates all auxiliary tables}
\usage{
UpdateAuxiliaryTables(shinyDir = ".")
}
\arguments{
\item{shinyDir}{The path of the shiny app directory.}
}
\description{
For performance, there are many auxiliary tables of pre-computed queries. This function rebuild all auxiliary tables and their indexes
}
