% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-Manipulation.R
\name{WriteLLTunaBilfishTbl}
\alias{WriteLLTunaBilfishTbl}
\title{Writes Longline Data to Database.}
\usage{
WriteLLTunaBilfishTbl(shinyDir, LL_tuna_bill_table)
}
\arguments{
\item{shinyDir}{The path of the shiny App directory.}

\item{LL_tuna_bill_table}{The data frame to be written on the database.}
}
\value{
TRUE if data table was written.
}
\description{
Takes a data frame containing longline data for tuna and billfish and writes it to disk.
}
